+++ src/main/java/com/sparta/levelup_backend/domain/alert/dto/AlertMessage.java
@@ -1,4 +1,4 @@
-package com.sparta.levelup_backend.domain.sse.dto;
+package com.sparta.levelup_backend.domain.alert.dto;
 
 public class AlertMessage {
 	public static final String ALREADY_MARKED_READ = "모든 메시지가 읽은 상태입니다.";
