+++ src/main/java/com/sparta/levelup_backend/domain/alert/repository/AlertMessageLogRepository.java
@@ -1,4 +1,4 @@
-package com.sparta.levelup_backend.domain.sse.repository;
+package com.sparta.levelup_backend.domain.alert.repository;
 
 import static com.sparta.levelup_backend.exception.common.ErrorCode.*;
 
@@ -6,8 +6,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.sparta.levelup_backend.domain.sse.dto.response.AlertLogResponseDto;
-import com.sparta.levelup_backend.domain.sse.entity.AlertMessageLogEntity;
+import com.sparta.levelup_backend.domain.alert.dto.response.AlertLogResponseDto;
+import com.sparta.levelup_backend.domain.alert.entity.AlertMessageLogEntity;
 import com.sparta.levelup_backend.exception.common.NotFoundException;
 
 public interface AlertMessageLogRepository extends JpaRepository<AlertMessageLogEntity, Long> {
