+++ src/main/java/com/sparta/levelup_backend/domain/user/dto/UserMessage.java
@@ -2,7 +2,7 @@ package com.sparta.levelup_backend.domain.user.dto;
 
 public class UserMessage {
 	public static final String PASSWORD_RESET_PREFIX = "인증 코드는 다음과 같습니다. \n 5분 안에 입력해주세요. \n 인증코드: ";
-	public static final String PASSWORD_RESET_SUBJECT = "Levelup Project 비밀번호 초기화 인증코드입니다.";
+	public static final String PASSWORD_RESET_SUBJECT = "Levelup 비밀번호 초기화 인증코드입니다.";
 	public static final String PASSWORD_RESET_CODE_PREFIX = "RESETCODE_";
-	public static final String CONGRATULATION_SIGNUP = "Levelup Project에 가입하신 고객님을 환영합니다.";
+	public static final String CONGRATULATION_SIGNUP = "Levelup에 가입하신 고객님을 환영합니다.";
 }
